+++ controllers/IndexController.js
@@ -1,19 +1,14 @@
-const { check, validationResult, body } = require("express-validator")
-
 const { User, School, Subject, Student, Teacher, Guardian, Class, Course, Student_Guardian, Class_Student, Lesson, Attendance, Evaluation, Student_Evaluation, Repeater } = require("../models");
 const bcrypt = require('bcrypt');
 const saltRounds = 10;
 
 module.exports = {
-    uploadFile: (req, res) => {
-        return res.render("teacher-home", { userPicture: user.picture })
-    },
     renderLogin: (req, res) => {
         return res.render('index');
     },
     login: async (req, res) => {
         // READ INFOS FROM REQ.BODY
-        let { userType, email, password } = req.body;
+        const { userType, email, password } = req.body;
         // TRY AND LOAD A USER FROM DB WHOSE EMAIL == REQ.BODY.EMAIL
         const user = await User.findOne(
             {
@@ -43,311 +38,9 @@ module.exports = {
             return res.redirect(`/responsavel/home`);
         }
     },
-    renderTeacherHome: async (req, res) => {
-        const user = req.session.user;
-
-        const teacher = await Teacher.findOne({ where: { userId: user.id } });
-        const courses = await Course.findAll({ where: { teacherId: teacher.id } });
-
-        const subjectsIds = courses.filter(course => course.subjectId);
-        const classesIds = courses.filter(course => course.classId);
-        const schoolsIds = courses.filter(course => course.schoolId);
-
-        let subjects = [];
-        for (id of subjectsIds) {
-            subjects.push(await Subject.findOne({ where: { id } }));
-        }
-        let classes = [];
-        for (id of classesIds) {
-            classes.push(await Subject.findOne({ where: { id } }));
-        }
-        let schools = [];
-        for (id of schoolsIds) {
-            schools.push(await Subject.findOne({ where: { id } }));
-        }
-
-        console.log(user.picture);
-
-        // AND THEN...
-        return res.render("teacher-home", { user, schools, classes });
-    },
-    renderGuardianHome: async (req, res) => {
-        // GET GUARDIAN DATA FROM DB,
-        // AND THEN...
-        return res.render("guardian-home");
-    },
-    redirectToRegistrationForm: (req, res) => {
+    register: (req, res) => {
         return req.query.usuario == "professor" ?
             res.redirect("/professor/cadastrar") :
             res.redirect("/responsavel/cadastrar");
-    },
-    renderTeacherRegistrationForm: (req, res) => {
-        res.render("register-teacher");
-    },
-    renderGuardianRegistrationForm: (req, res) => {
-        res.render("register-guardian");
-    },
-    registerTeacher: async (req, res) => {
-
-        // CREATE USER
-        const { forename, surname, email, phone, password } = req.body;
-        const picture = req.file ? req.file.filename : null;
-
-        const user = await User.create(
-            {
-                forename: forename.toUpperCase(),
-                surname: surname.toUpperCase(),
-                email: email.toUpperCase(),
-                phone,
-                password: bcrypt.hashSync(password, saltRounds),
-                picture
-            }
-        );
-        // MAKE THE USER A TEACHER
-        const teacher = await Teacher.create({ userId: user.id });
-
-        // CREATE SCHOOLS
-        const objKeysSchool = Object.keys(req.body).filter(
-            key => key.substr(0, 6) == "school"
-        );
-        let schoolsList = [];
-        for (schoolKey of objKeysSchool) {
-            schoolsList.push(
-                {
-                    name: req.body[schoolKey][2],
-                    passingGrade: req.body[schoolKey][3],
-                    academicTerms: req.body[schoolKey][4],
-                    state: req.body[schoolKey][0],
-                    municipality: req.body[schoolKey][1]
-                }
-            );
-        };
-        const schools = await School.bulkCreate(schoolsList);
-
-        // CREATE CLASSES
-        const objKeysClass = Object.keys(req.body).filter( // ex. return [ "class1-school1", "class2-school1", "class1-school2" ]
-            key => key.substr(0, 5) == "class"
-        );
-        let classesList = [];
-        for (let i = 0; i < schools.length; i++) {
-            let thisSchoolClasses = objKeysClass.filter(
-                thisClass => thisClass.includes(`school${i + 1}`)
-            );
-            for (aClass of thisSchoolClasses) {
-                let lvl = req.body[aClass][1].split("-"); // ex. return [ "Ensino Fundamental ", " 6º ano" ]
-                classesList.push(
-                    {
-                        schoolId: schools[i].id, // ASSOCIATE CLASSES TO A SCHOOL
-                        code: req.body[aClass][2],
-                        year: req.body[aClass][0],
-                        levelOfEducation: lvl[0].trim(),
-                        grade: lvl[1].trim()
-                    }
-                );
-            };
-        };
-        const classes = await Class.bulkCreate(classesList);
-
-        // CREATE COURSES
-        const objKeysSubjects = Object.keys(req.body).filter( // ex. return [ "subjects-class1-school1", "subjects-class2-school1", "subjects-class1-school2" ]
-            key => key.substr(0, 8) == "subjects"
-        );
-        let coursesList = [];
-        for (let i = 0; i < classes.length; i++) {
-            let thisClassSubjectsKey = objKeysSubjects.filter(
-                subjects => subjects.includes(`class${i + 1}`)
-            );
-            let thisClassSubjects = req.body[thisClassSubjectsKey]; // array containing the subjects ids of only one class
-
-            for (subjectId of thisClassSubjects) {
-                let course = await Course.create(
-                    {
-                        teacherId: teacher.id,
-                        subjectId,
-                        classId: classes[i].id
-                    }
-                );
-                coursesList.push(course);
-            }
-        }
-
-        // CREATE STUDENTS
-        const objKeysStudent = Object.keys(req.body).filter( // ex. return [ "student1-class1-school1", "student2-class1-school1", "student1-class2-school1" ]
-            key => key.substr(0, 7) == "student"
-        );
-        for (let i = 0; i < classes.length; i++) {
-            let thisClassStudentsKeys = objKeysStudent.filter(
-                student => student.includes(objKeysClass[i])
-            );
-            let thisClassStudents = []; // array of arrays like ["student class number", "student name", "checkbox on", "retaken course subjectId"]
-            for (key of thisClassStudentsKeys) {
-                thisClassStudents.push(req.body[key]);
-            }
-            for (student of thisClassStudents) { // array of arrays
-
-                // VALIDATE STUDENTS
-                let registeredStudent = await Student.findOne({ where: { name: student[1].trim() } });
-                if (registeredStudent) {
-                    return res.render("register-teacher", { errors: ["Cadastre novos alunos."] })
-                }
-
-                // CREATE STUDENTS
-                let newStudent = await Student.create(
-                    {
-                        name: student[1].trim()
-                    }
-                );
-
-                // ASSOCIATE STUDENT TO CLASS
-                await Class_Student.create(
-                    {
-                        classId: classes[i].id,
-                        studentId: newStudent.id,
-                        number: student[0]
-                    }
-                );
-
-                // CREATE REPEATER
-                let repeater = false;
-                if (student[2] == "on") { repeater = true; } // checkbox checked
-
-                if (repeater) {
-                    let thisClassCourses = coursesList.filter(
-                        course => course.classId == classes[i].id
-                    );
-                    let coursesIdsList = [];
-
-                    for (let j = 3; j < student.length; j++) { // from student[3], there's a list of subjects ids the student is repeating
-
-                        for (course of thisClassCourses) {
-                            if (course.subjectId == student[j]) {
-                                coursesIdsList.push(course.id);
-                            }
-                        }
-                    }
-                    for (courseId of coursesIdsList) {
-                        await Repeater.create(
-                            {
-                                studentId: newStudent.id,
-                                courseId
-                            }
-                        );
-                    }
-                }
-            }
-        }
-        // SET A SESSION FOR THE USER
-        req.session.user = user;
-        // AND REDIRECT HOME
-        return res.redirect("/professor/home");
-    },
-    registerGuardian: async (req, res, next) => {
-
-        // CREATE USER 
-        const { forename, surname, email, phone, password } = req.bory;
-        let picture = req.file ? req.file.name : null;
-
-        let invalidEmail = await User.findOne({ where: { email }});
-        if (invalidEmail) {
-            return res.send(`O email ${email} já está cadastrado!`)
-        }
-        const user = User.create({
-            forename,
-            surname,
-            email,
-            phone,
-            password: bcrypt.hashSync(password, saltRounds),
-            picture
-        });
-
-        // MAKE THE USER A GUARDIAN
-        const guardian = await Guardian.create({userId: user.id});
-
-        const { selectState, selectCity } = req.body;
-        
-        // SCHOOL LIST
-        const schools = await School.findAll({ where: { state: selectState, municipality: selectCity }});
-        const school = req.body.selectSchool;
-        
-        // CLASS LIST
-        const classes = await Class.findAll({ where: { schoolId: school.id }});
-        const clas = req.body.selectClass;
-        
-        const classStudent = await Class_Student.findAll({ where: { classId: clas.id}});
-
-        // STUDENT LIST
-        const students = await Student.findAll({ where: { id: classStudent.studentId }});
-        const selectStudent = req.body.selectStudent;
-        const student = await Student.findOne({ where: { name: selectStudent }});
-       
-        // ASSOCIATE STUDENT TO GUARDIAN
-        const student_guardian = await Student_Guardian.create({studentId: student.id, guardianId: guardian.id});
-
-        // SET A SESSION
-        req.session.user = user;
-
-        return res.redirect("/responsavel/home");
-    },
-    renderTeacherUpdateForm: async (req, res) => {
-        // LOAD USER FROM DB
-        // PASS OBJECT USER INTO RENDER METHOD
-        return res.render("update-teacher");
-    },
-    renderGuardianUpdateForm: async (req, res) => {
-
-    },
-    updateTeacher: async (req, res, next) => {
-        // GET REQ.BODY CONTENT
-        // AND UPDATE DATA IN DB
-        // const result = await User.update({
-        //     // DATA TO UPDATE
-        // },
-        // {
-        //     where: {
-        //         id
-        //     }
-        // });
-    },
-    updateGuardian: async (req, res, next) => {
-
-    },
-    deleteTeacher: async (req, res) => {
-        // RETRIEVE USER ID,
-        // AND DELETE ONLY TEACHER USER TYPE
-        // OR DELETE USER FROM DB
-        // const result = await User.destroy(
-        //     {
-        //         where: {
-
-        //         }
-        //     }
-        // );
-    },
-    deleteGuardian: async (req, res) => {
-        // RETRIEVE USER ID,
-        // AND DELETE ONLY GUARDIAN USER TYPE
-        // OR DELETE USER FROM DB
-        // const result = await User.destroy(
-        //     {
-        //         where: {
-
-        //         }
-        //     }
-        // );
-    },
-    renderGradeBook: (req, res) => {
-        return res.render("set-notes");
-    },
-    recordGrades: (req, res) => {
-
-    },
-    renderAttendanceSheet: (req, res) => {
-        return res.render("attendance");
-    },
-    recordAttendances: (req, res) => {
-
-    },
-    renderRecordBook: (req, res) => {
-        return res.render('daily');
     }
 };
