--- controllers/UsersController.js
@@ -1,5 +1,0 @@
-module.exports = {
-  create: (req, res) => {
-    res.render("create");
-  },
-};
