+++ views/index.ejs
@@ -11,7 +11,7 @@
 
   <section id="about">
     <article>
-      <h2>PROFESSOR,</h2>
+      <h2>PROFESSORES,</h2>
       <p>Este é o seu DIÁRIO ESCOLAR virtual!</p>
       <ul>
         <li>Faça chamadas,</li>
@@ -20,8 +20,8 @@
         <li>contate responsáveis.</li>
       </ul>
       <br>
-      <h2>RESPONSÁVEL,</h2>
-      <p>Acompanhe o desempenho escolar dos seus filhos!</p>
+      <h2>RESPONSÁVEIS,</h2>
+      <p>Acompanhem o desempenho escolar dos seus filhos!</p>
       <p>
         Receba avisos direto dos professores e não perca a data da próxima
         reunião de Pais e Mestres.
@@ -34,7 +34,7 @@
 
       <form action="/cadastrar" method="GET" id="signup">
         <div id="user-type">
-          <label for="teacher" class="selected">Sou PROFESSOR<input type="radio" name="usuario" value="professor"
+          <label for="teacher" class="selected">Sou DOCENTE<input type="radio" name="usuario" value="professor"
               id="teacher" checked /></label>
           <label for="guardian">Sou RESPONSÁVEL<input type="radio" name="usuario" value="responsavel"
               id="guardian" /></label>
