+++ views/teacher-home.ejs
@@ -16,7 +16,7 @@
                 <img src="../<%= user.picture %>" alt="avatar">
                 <span>
                     <p><%= user.forename + " " + user.surname %></p>
-                    <p>PROFESSORX</p>
+                    <p>DOCENTE</p>
                 </span>
             </div>
         </div>
